requirements.txt
beautifulsoup4>=4.14.2
feedparser>=6.0.12
google-generativeai>=0.8.0
html-to-markdown>=2.7.1
markdown>=3.7.0
markdownify>=0.11.6
pydantic>=2.0.0
python-dotenv>=1.2.1
requests>=2.32.5
sqlalchemy>=2.0.44
youtube-transcript-api>=1.2.3
yt-dlp>=2025.11.12

